+++ .gitignore
@@ -14,3 +14,5 @@ TS.log
 *.sln
 ptvsd/*
 .vs/*
+sublime.py
+sublime_plugin.py
